🐙ueli🎆26🌙06☀️04🚩308
🎆26🌙07☀️19⏰03⌚25⏱️19
🧮Scaffold mathematical graph NetworkX-parity crates for paths, flow, spectral, and related algorithms
⚡Slim energy engine HVAC and zone modules toward absolute feature completeness
🏛️Trim architect program validation/trace and shrink norm plugin Eurocode session surface
📚Overhaul storybook scopes with OS stories and multi-scope wiring in root script.ts
🖥️Refresh ui wgpu and react packages for apps-compose framework-owned UI
📦Register new mathematical graph and random crates in workspace Cargo.toml and project.json
🎞️Touch animate core text/animation helpers and fem 2d/3d app surfaces
🛠️Add repo cli project.json and compose/cad/coda script tooling for leveled tests

📊uloc➕3598✏️1522➖5521🟰10641
🧾410k➕1867✏️5➖10🟰1882
🦀305k➕385✏️1487➖5510🟰7382
🟦162k➕715✏️30➖1🟰746
🔵95k
🟣33k
🐍31k
🎨21k
🟨16k
📝12k
⚙️6k➕631🟰631
🛢️3k
🌐3k
📋1k
💠893
🐚586
📑57
🐳43

+++ animate/core/rs/src/axes.rs
@@ -143,8 +143,8 @@ impl NumberPlane {
             Box::new(arrow(origin, Point::new(origin.x(), origin.y() + y_len), color, 3.0, 0.2)),
         ];
         let grid_color = color.with_alpha(0.25);
-        let x_steps = ((x_range.1 - x_range.0) as i32).abs().max(1) as i32;
-        let y_steps = ((y_range.1 - y_range.0) as i32).abs().max(1) as i32;
+        let x_steps = ((x_range.1 - x_range.0) as i32).abs().max(1);
+        let y_steps = ((y_range.1 - y_range.0) as i32).abs().max(1);
         for i in 0..=x_steps {
             let x = origin.x() + i as f64 * unit_size;
             children.push(Box::new(line(
